correctly generating all tables for 'local' now
