+++ scripts/data_generators/connections/__init__.py
@@ -1,23 +1,16 @@
-from typing import Dict, Type
+import os
+import pkgutil
+import importlib
 
 
-class IcebergConnection:
-    registry: Dict[str, Type['IcebergConnection']] = {}
+# --- Dynamic importing logic to auto-register all connections ---
+def _import_all_submodules():
+    package_dir = os.path.dirname(__file__)
+    for finder, name, ispkg in pkgutil.iter_modules([package_dir]):
+        if ispkg:
+            importlib.import_module(f".{name}", __name__)
 
-    def __init__(self, name: str):
-        self.name = name
 
-    @classmethod
-    def register(cls, name: str):
-        def decorator(subclass):
-            cls.registry[name] = subclass
-            return subclass
+_import_all_submodules()
 
-        return decorator
-
-    @classmethod
-    def get_class(cls, name: str):
-        if name not in registry:
-            print(f"'{name}' not in registry, exiting!")
-            exit(1)
-        return registry[name]
+from scripts.data_generators.connections.base import IcebergConnection
