+++ scripts/data_generators/connections/lakekeeper/__init__.py
@@ -3,7 +3,7 @@ import pyspark
 import pyspark.sql
 from pyspark import SparkContext
 
-from . import IcebergConnection
+from ..base import IcebergConnection
 
 import sys
 import os
@@ -30,8 +30,8 @@ ICEBERG_VERSION = "1.7.0"
 @IcebergConnection.register(CONNECTION_KEY)
 class IcebergSparkLocal(IcebergConnection):
     def __init__(self):
-        super().__init__(CONNECTION_KEY)
-        self.con = get_connection(self)
+        super().__init__(CONNECTION_KEY, 'lakekeeper')
+        self.con = self.get_connection()
 
     def get_connection(self):
         conf = {
