+++ scripts/data_generators/tests/pyspark_iceberg_table_v1/__init__.py
@@ -1,6 +1,6 @@
 import duckdb
 import os
-from .. import IcebergTest
+from scripts.data_generators.tests.base import IcebergTest
 import pathlib
 import tempfile
 
@@ -47,4 +47,4 @@ class Test(IcebergTest):
         duckdb_con.execute(f"COPY test_table TO '{self.parquet_file}' (FORMAT PARQUET)")
 
     def setup(self, con):
-        con.read.parquet(self.parquet_file).createOrReplaceTempView('parquet_file_view')
+        con.con.read.parquet(self.parquet_file.as_posix()).createOrReplaceTempView('parquet_file_view')
