+++ scripts/data_generators/tests/quickstart_table/q00.sql
@@ -1,5 +1,4 @@
-DROP TABLE IF EXISTS quickstart_table;
-CREATE TABLE IF NOT EXISTS quickstart_table (
+CREATE OR REPLACE TABLE quickstart_table (
      id BIGINT, data STRING
 )
 USING ICEBERG;
