+++ scripts/data_generators/tests/table_with_deletes/q01.sql
@@ -1,4 +1,4 @@
-update iceberg_catalog.table_with_deletes
+update default.table_with_deletes
 set l_orderkey=NULL,
     l_partkey=NULL,
     l_suppkey=NULL,
